pyproject.toml
[project]
name = "aseoka-hub"
version = "1.0.0"
description = "ASEOKA Hub - Central coordination server for SEO agents"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "ASEOKA Team" }]
keywords = ["seo", "agent", "hub", "server"]

dependencies = [
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",

    # Database
    "aiosqlite>=0.19.0",

    # Authentication & Security
    "bcrypt>=4.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",

    # Config & Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # HTTP Client
    "httpx>=0.27.0",

    # CLI
    "typer>=0.9.0",
    "rich>=13.7.0",

    # Logging
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[project.scripts]
aseoka-hub = "aseoka_hub.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aseoka_hub"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
